moved profile page to pages folder and fetch information to populate profile section
