+++ src/pages/Profile.jsx
@@ -1,5 +1,5 @@
-import ShopperInventory from "./components/ProfilePage/ShopperInventory.jsx";
-import ShopperProfile from "./components/ProfilePage/ShopperProfile.jsx";
+import ShopperInventory from "../components/ProfilePage/ShopperInventory.jsx";
+import ShopperProfile from "../components/ProfilePage/ShopperProfile.jsx";
 
 function ProfilePage() {
